requirements.txt
requests~=2.31.0
docx~=0.2.4
python-dateutil~=2.8.2